Move context logic to new class, with new helper methods
